+++ exercises/leap/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn is_leap_year(year: i32) -> bool {
+pub fn is_leap_year(year: u64) -> bool {
     unimplemented!("true if {} is a leap year", year)
 }
